Moved is_onefi_ssid function to hooks folder
