+++ src/screens/ConnectScreen.tsx
@@ -3,7 +3,6 @@ import { useCallback, useEffect, useState } from "react";
 import {
   View,
   Text,
-  ImageBackground,
   FlatList,
   TouchableOpacity,
   StyleSheet,
@@ -28,23 +27,7 @@ import WifiLevelIcon from "../icons/WifiLevelIcon";
 import LockICon from "../icons/LockIcon";
 import { globalStyle } from "../constants/globalStyle";
 
-//import { is_onefi_ssid } from "../../api/client";
-/**
- * Check if the given SSID matches the OneFi format.
- * @param {string} ssid - SSID candidate
- * @returns {boolean} true: matches, false: doesn't match.
- */
-function is_onefi_ssid(ssid: string) {
-  if (ssid.length > 2) {
-    if (ssid[0] === "O" && ssid[1] === "F") {
-      let base64_part = ssid.substring(2);
-      let decoded_str = Buffer.from(base64_part, "base64").toString("hex");
-      return decoded_str.length === 42;
-    }
-  }
-
-  return false;
-}
+import { is_onefi_ssid } from "../hooks/is_onefi_ssid";
 
 async function getPermission() {
   if (Platform.OS === "android") {
